auto-engineer/requirements.txt
PyGithub==2.2.0
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.8.2
